[elasticsearch-utils] Move ES test class and remove module.
