--- heroic-elasticsearch-test/build.gradle
@@ -1,10 +1,0 @@
-dependencies {
-    implementation project(':heroic-test-it')
-    implementation project(':heroic-elasticsearch-utils')
-    implementation project(':heroic-component')
-    implementation project(path: ':heroic-component', configuration: 'testRuntime')
-
-    implementation 'com.google.guava:guava'
-}
-
-description = 'Heroic: Elasticsearch Test Project'
