+++ heroic-elasticsearch-utils/build.gradle
@@ -2,6 +2,8 @@ dependencies {
     implementation project(':heroic-component')
     implementation 'com.spotify.heroic.repackaged:folsom'
     api 'org.elasticsearch.client:transport:5.3.2'
+    
+    implementation project(path: ':heroic-component', configuration: 'testRuntime')
 }
 
 description = 'Heroic: Elasticsearch Utilities'
